+++ ext/tensor.c
@@ -24,8 +24,10 @@
 
 
 zend_class_entry *tensor_arraylike_ce;
-zend_class_entry *tensor_decompositions_decomposition_ce;
 zend_class_entry *tensor_tensor_ce;
+zend_class_entry *tensor_decompositions_decomposition_ce;
+zend_class_entry *tensor_statistical_ce;
+zend_class_entry *tensor_trigonometric_ce;
 zend_class_entry *tensor_vector_ce;
 zend_class_entry *tensor_columnvector_ce;
 zend_class_entry *tensor_decompositions_lu_ce;
@@ -44,8 +46,10 @@ static PHP_MINIT_FUNCTION(tensor)
 	REGISTER_INI_ENTRIES();
 	zephir_module_init();
 	ZEPHIR_INIT(Tensor_ArrayLike);
-	ZEPHIR_INIT(Tensor_Decompositions_Decomposition);
 	ZEPHIR_INIT(Tensor_Tensor);
+	ZEPHIR_INIT(Tensor_Decompositions_Decomposition);
+	ZEPHIR_INIT(Tensor_Statistical);
+	ZEPHIR_INIT(Tensor_Trigonometric);
 	ZEPHIR_INIT(Tensor_Vector);
 	ZEPHIR_INIT(Tensor_ColumnVector);
 	ZEPHIR_INIT(Tensor_Decompositions_Lu);