+++ ext/tensor.h
@@ -5,8 +5,10 @@
 #define ZEPHIR_CLASS_ENTRIES_H
 
 #include "tensor/arraylike.zep.h"
-#include "tensor/decompositions/decomposition.zep.h"
 #include "tensor/tensor.zep.h"
+#include "tensor/decompositions/decomposition.zep.h"
+#include "tensor/statistical.zep.h"
+#include "tensor/trigonometric.zep.h"
 #include "tensor/vector.zep.h"
 #include "tensor/columnvector.zep.h"
 #include "tensor/decompositions/lu.zep.h"