+++ ext/tensor/vector.zep.c
@@ -63,6 +63,8 @@ ZEPHIR_INIT_CLASS(Tensor_Vector) {
 	zephir_declare_class_constant_double(tensor_vector_ce, SL("M_E"), 2.7182818284590452354);
 
 	zend_class_implements(tensor_vector_ce, 1, tensor_tensor_ce);
+	zend_class_implements(tensor_vector_ce, 1, tensor_trigonometric_ce);
+	zend_class_implements(tensor_vector_ce, 1, tensor_statistical_ce);
 	return SUCCESS;
 
 }