+++ tests/VectorTest.php
@@ -6,7 +6,9 @@ use Tensor\Tensor;
 use Tensor\Vector;
 use Tensor\Matrix;
 use Tensor\ArrayLike;
+use Tensor\Statistical;
 use Tensor\ColumnVector;
+use Tensor\Trigonometric;
 use PHPUnit\Framework\TestCase;
 use InvalidArgumentException;
 use IteratorAggregate;
@@ -42,6 +44,8 @@ class VectorTest extends TestCase
     public function test_build_vector()
     {
         $this->assertInstanceOf(Vector::class, $this->a);
+        $this->assertInstanceOf(Trigonometric::class, $this->a);
+        $this->assertInstanceOf(Statistical::class, $this->a);
         $this->assertInstanceOf(Tensor::class, $this->a);
         $this->assertInstanceOf(ArrayLike::class, $this->a);
     }