+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     - Added return sub matrix
     - Added positive definite and semidefinite methods to Matrix
     - Added is symmetric method to Matrix
+    - Added Statistical and Trigonometric interfaces
 
 - 1.0.4
     - Added transpose to Tensor API